Refactor Record Transaction from Penjualan andPembelian -> Pemasukkan dan Pengeluaran by 1. Deleting FeeType and Fee on both sides 2. Rename the package (add: balance_sheet also) 3. Refactor the MainMenu for Report -> update the names of each items 4. Add accountName with AsetLancar, AsetTetap, and Hutang on strings.xml 5. Add purchaseType on strings.xml with HPP, AsetTetap, BebanBeban
