+++ app/src/main/java/id/novian/flowablecash/view/report/main/MainReport.kt
@@ -46,9 +46,9 @@ class MainReport : BaseFragment<FragmentMainReportBinding>() {
 
     private fun onClickOnItem(journalType: String) {
         when (journalType) {
-            "Jurnal Penjualan" -> moveTo(R.id.action_mainReport_to_cashReceiptJournal)
-            "Jurnal Pembelian" -> moveTo(R.id.action_mainReport_to_purchasesJournal)
-            "Neraca Saldo" -> moveTo(R.id.action_mainReport_to_balanceSheetFragment)
+            "Rekap Pemasukkan" -> moveTo(R.id.action_mainReport_to_cashReceiptJournal)
+            "Rekap Pengeluaran" -> moveTo(R.id.action_mainReport_to_purchasesJournal)
+            "Rekap Harga Pokok Penjualan" -> moveTo(R.id.action_mainReport_to_balanceSheetFragment)
             "Laporan Laba Rugi" -> moveTo(R.id.action_mainReport_to_incomeStatements)
             "Laporan Posisi Keuangan" -> moveTo(R.id.action_mainReport_to_financialPosition)
         }
