+++ app/src/main/java/id/novian/flowablecash/view/report/main/ModelMenu.kt
@@ -8,9 +8,9 @@ data class ModelMenu(
 )
 
 val menuReportItems = listOf(
-    ModelMenu("Jurnal Penjualan", R.drawable.ic_cash_receipt),
-    ModelMenu("Jurnal Pembelian", R.drawable.ic_purchases_journal),
-    ModelMenu("Neraca Saldo", R.drawable.ic_balance_sheet),
+    ModelMenu("Rekap Pemasukkan", R.drawable.ic_cash_receipt),
+    ModelMenu("Rekap Pengeluaran", R.drawable.ic_purchases_journal),
+    ModelMenu("Rekap Harga Pokok Penjualan", R.drawable.ic_balance_sheet),
     ModelMenu("Laporan Laba Rugi", R.drawable.ic_income_statements),
     ModelMenu("Laporan Posisi Keuangan", R.drawable.ic_financial_position)
 )
