+++ app/src/main/java/id/novian/flowablecash/view/report/rekap_hpp/TrialBalanceAdapter.kt
@@ -1,4 +1,4 @@
-package id.novian.flowablecash.view.report.balance_sheet
+package id.novian.flowablecash.view.report.rekap_hpp
 
 import android.graphics.Typeface
 import android.view.Gravity
