+++ app/src/main/java/id/novian/flowablecash/view/report/rekap_pemasukkan/CashReceiptJournal.kt
@@ -1,4 +1,4 @@
-package id.novian.flowablecash.view.report.cash_receipt
+package id.novian.flowablecash.view.report.rekap_pemasukkan
 
 import android.util.Log
 import android.view.LayoutInflater
