+++ app/src/main/java/id/novian/flowablecash/view/report/rekap_pemasukkan/CashReceiptJournalViewModel.kt
@@ -1,48 +1,55 @@
-package id.novian.flowablecash.view.report.purchases_journal
+package id.novian.flowablecash.view.report.rekap_pemasukkan
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import com.google.gson.Gson
 import dagger.hilt.android.lifecycle.HiltViewModel
 import id.novian.flowablecash.base.vm.BaseViewModel
-import id.novian.flowablecash.domain.models.PurchasesJournal
-import id.novian.flowablecash.domain.repository.PurchasesJournalRepository
+import id.novian.flowablecash.domain.models.CashReceiptJournal
+import id.novian.flowablecash.domain.repository.CashReceiptJournalRepository
 import id.novian.flowablecash.helpers.CreateToast
 import id.novian.flowablecash.helpers.Result
+import id.novian.flowablecash.usecase.posting.PostingUseCase
 import io.reactivex.rxjava3.core.Scheduler
+import java.util.Calendar
 import javax.inject.Inject
 import javax.inject.Named
 
 @HiltViewModel
-class PurchasesJournalViewModel @Inject constructor(
-    private val repo: PurchasesJournalRepository,
+class CashReceiptJournalViewModel @Inject constructor(
     @Named("IO") private val schedulerIo: Scheduler,
     @Named("MAIN") private val schedulerMain: Scheduler,
-    private val toast: CreateToast
+    private val repo: CashReceiptJournalRepository,
+    private val toast: CreateToast,
+    private val gson: Gson,
+    private val postingUseCase: PostingUseCase,
+    private val calendar: Calendar
 ) : BaseViewModel() {
-    private val _dataPurchaseJournal: MutableLiveData<List<PurchasesJournal>> = MutableLiveData()
-    val dataPurchaseJournal: LiveData<List<PurchasesJournal>> get() = _dataPurchaseJournal
+    private val _dataCashReceiptJournal: MutableLiveData<List<CashReceiptJournal>> =
+        MutableLiveData()
+    val dataCashReceiptJournal: LiveData<List<CashReceiptJournal>> get() = _dataCashReceiptJournal
 
     private val _result: MutableLiveData<Result> = MutableLiveData()
     val result: LiveData<Result> get() = _result
 
-    private val observablePurchasesJournal = repo.getJournal()
+    private val observableCashReceiptJournal = repo.getJournal()
         .share()
 
-    fun getJournal() {
-        val disposable = observablePurchasesJournal
+    fun getCashReceiptJournal() {
+        val disposable = observableCashReceiptJournal
             .subscribeOn(schedulerIo)
             .observeOn(schedulerMain)
             .doOnSubscribe { _result.postValue(Result.LOADING) }
             .subscribe({
-                _dataPurchaseJournal.postValue(it)
                 _result.postValue(Result.SUCCESS)
+                _dataCashReceiptJournal.postValue(it)
             }, {
                 it.printStackTrace()
-                errorMessage.postValue(it.message)
                 _result.postValue(Result.FAILED)
+                errorMessage.postValue(it.message)
             })
 
-        compositeDisposable.add(disposable)
+        compositeDisposable.addAll(disposable)
     }
 
     fun createToast(message: String) {
