+++ app/src/main/java/id/novian/flowablecash/view/report/rekap_pengeluaran/PurchasesJournalViewModel.kt
@@ -1,55 +1,48 @@
-package id.novian.flowablecash.view.report.cash_receipt
+package id.novian.flowablecash.view.report.rekap_pengeluaran
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.google.gson.Gson
 import dagger.hilt.android.lifecycle.HiltViewModel
 import id.novian.flowablecash.base.vm.BaseViewModel
-import id.novian.flowablecash.domain.models.CashReceiptJournal
-import id.novian.flowablecash.domain.repository.CashReceiptJournalRepository
+import id.novian.flowablecash.domain.models.PurchasesJournal
+import id.novian.flowablecash.domain.repository.PurchasesJournalRepository
 import id.novian.flowablecash.helpers.CreateToast
 import id.novian.flowablecash.helpers.Result
-import id.novian.flowablecash.usecase.posting.PostingUseCase
 import io.reactivex.rxjava3.core.Scheduler
-import java.util.Calendar
 import javax.inject.Inject
 import javax.inject.Named
 
 @HiltViewModel
-class CashReceiptJournalViewModel @Inject constructor(
+class PurchasesJournalViewModel @Inject constructor(
+    private val repo: PurchasesJournalRepository,
     @Named("IO") private val schedulerIo: Scheduler,
     @Named("MAIN") private val schedulerMain: Scheduler,
-    private val repo: CashReceiptJournalRepository,
-    private val toast: CreateToast,
-    private val gson: Gson,
-    private val postingUseCase: PostingUseCase,
-    private val calendar: Calendar
+    private val toast: CreateToast
 ) : BaseViewModel() {
-    private val _dataCashReceiptJournal: MutableLiveData<List<CashReceiptJournal>> =
-        MutableLiveData()
-    val dataCashReceiptJournal: LiveData<List<CashReceiptJournal>> get() = _dataCashReceiptJournal
+    private val _dataPurchaseJournal: MutableLiveData<List<PurchasesJournal>> = MutableLiveData()
+    val dataPurchaseJournal: LiveData<List<PurchasesJournal>> get() = _dataPurchaseJournal
 
     private val _result: MutableLiveData<Result> = MutableLiveData()
     val result: LiveData<Result> get() = _result
 
-    private val observableCashReceiptJournal = repo.getJournal()
+    private val observablePurchasesJournal = repo.getJournal()
         .share()
 
-    fun getCashReceiptJournal() {
-        val disposable = observableCashReceiptJournal
+    fun getJournal() {
+        val disposable = observablePurchasesJournal
             .subscribeOn(schedulerIo)
             .observeOn(schedulerMain)
             .doOnSubscribe { _result.postValue(Result.LOADING) }
             .subscribe({
+                _dataPurchaseJournal.postValue(it)
                 _result.postValue(Result.SUCCESS)
-                _dataCashReceiptJournal.postValue(it)
             }, {
                 it.printStackTrace()
-                _result.postValue(Result.FAILED)
                 errorMessage.postValue(it.message)
+                _result.postValue(Result.FAILED)
             })
 
-        compositeDisposable.addAll(disposable)
+        compositeDisposable.add(disposable)
     }
 
     fun createToast(message: String) {
