+++ app/src/main/res/layout/fragment_purchases_journal.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".view.report.purchases_journal.PurchasesJournal">
+    tools:context=".view.report.rekap_pengeluaran.PurchasesJournal">
 
     <ImageView
         android:id="@+id/btn_back"
